Former-commit-id: 314a21f9e58f6c8a23fae1a920196b6bb73de206

+++ applications/plugins/SofaTest/SofaTest_test/BezierTetrahedronTopology_test.cpp
@@ -147,7 +147,7 @@ struct BezierTetrahedronTopology_test : public Sofa_test<typename _DataTypes::Re
 		BezierDegreeType degree=container->getDegree();
 		// check the total number of vertices.
 		size_t nbPoints=container->getNumberOfTetrahedralPoints()+container->getNumberOfEdges()*(degree-1)+container->getNumberOfTriangles()*(degree-1)*(degree-2)/2+container->getNumberOfTetrahedra()*((degree-1)*(degree-2)*(degree-3)/6);
-		if(container->getNbPoints()!=nbPoints) {
+        if((size_t)container->getNbPoints()!=nbPoints) {
 			ADD_FAILURE() << "wrong number of points " <<container->getNbPoints() << " is wrong. It should be  " <<nbPoints  << std::endl;
 			return false;
 		}
